%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{HaarSeg-package}
\alias{HaarSeg-package}
\alias{HaarSeg}
\docType{package}

\title{Package HaarSeg}


\description{
  A fast and flexible method for the segmentation of aCGH data using the HaarSeg method by Ben-Yaacov and Eldar (2008) <doi:10.1093/bioinformatics/btn272>..
}

\section{Dependancies and other requirements}{
  None.
}

\section{To get started}{
  \code{\link{haarSeg}}()
}

\references{
  [1] Ben-Yaacov E. and Eldar YC. \emph{A fast and flexible method for the segmentation of aCGH data}, Bioinformatics, 2008.
  \url{https://www.ee.technion.ac.il/Sites/People/YoninaEldar/Info/software/HaarSeg.htm}
}

\author{Erez Ben-Yaacov.  R package created by Henrik Bengtsson.}
\keyword{package}
