%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{HaarSeg-package}
\alias{HaarSeg-package}
\alias{HaarSeg}
\docType{package}

\title{Package HaarSeg}


\description{
  This package provides the HaarSeg method..
}

\section{Installation}{
  To install this package, please do follow the instructions at
  \url{http://www.braju.com/R/}.
}

\section{Dependancies and other requirements}{
  None.
}

\section{To get started}{
  \code{\link{haarSeg}}()
}

\section{How to cite this package}{
 \preformatted{
  Erez Ben-Yaacov and Yonina C. Eldar (R package wrapper by Henrik Bengtsson <henrikb@braju.com>) (2009). HaarSeg: HaarSeg. R package version 0.0.2. http://www.ee.technion.ac.il/Sites/People/YoninaEldar/Info/software/HaarSeg.htm
 }
}
\references{
  [1] Ben-Yaacov E. and Eldar YC. \emph{A fast and flexible method for the segmentation of aCGH data}, Bioinformatics, 2008.
  \url{http://www.ee.technion.ac.il/Sites/People/YoninaEldar/Info/software/HaarSeg.htm}
}

\author{Erez Ben-Yaacov.  R package created by Henrik Bengtsson.}
\keyword{package}
